+++ src/config/main.rs
@@ -56,8 +56,6 @@ fn default_log_level() -> LogLevel { LogLevel::Info }
 
 fn default_login_message() -> String { "Do not sign this message on other sites!".to_string() }
 
-fn default_post_character_limit() -> usize { 2000 }
-
 #[derive(Clone, Deserialize)]
 pub struct Config {
     // Properties auto-populated from the environment
@@ -99,8 +97,7 @@ pub struct Config {
     #[serde(default = "default_login_message")]
     pub login_message: String,
 
-    #[serde(default = "default_post_character_limit")]
-    pub post_character_limit: usize,
+    pub(super) post_character_limit: Option<usize>, // deprecated
 
     proxy_url: Option<String>,